--- src/pages/nonMbLookup/components/InputWrapper.styled.ts
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-
-export const InputWrapperLayout = styled.section`
-  display: flex;
-  flex-direction: column;
-  gap: 1.6rem;
-  padding: 2.4rem;
-`;
-
-export const EyeTest = styled.button`
-  position: relative;
-  right: 0;
-`;
